chore: move common code to base create order request/response classes
