+++ lib/binance_client/requests/base_create_order_request.rb
@@ -1,0 +1,25 @@
+module BinanceClient
+  class BaseCreateOrderRequest < AuthenticatedBaseRequest
+
+    attribute :symbol, String
+    attribute :side, String
+    attribute :type, String
+    attribute :quantity, String
+
+    private
+
+    def default_action
+      :post
+    end
+
+    def params_without_signature
+      {
+        symbol: symbol,
+        side: side,
+        type: type,
+        quantity: quantity,
+      }
+    end
+
+  end
+end
