+++ lib/binance_client/requests/create_test_order_request.rb
@@ -1,9 +1,5 @@
 module BinanceClient
-  class CreateTestOrderRequest < AuthenticatedBaseRequest
-    attribute :symbol, String
-    attribute :side, String
-    attribute :type, String
-    attribute :quantity, String
+  class CreateTestOrderRequest < BaseCreateOrderRequest
 
     private
 
@@ -11,17 +7,5 @@ module BinanceClient
       "/api/v3/order/test"
     end
 
-    def default_action
-      :post
-    end
-
-    def params_without_signature
-      {
-        symbol: symbol,
-        side: side,
-        type: type,
-        quantity: quantity,
-      }
-    end
   end
 end
